# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/requests": {
      "post": {
        "operationId": "RequestsController_create",
        "summary": "Create public request (no auth)",
        "description": "Creates a request and publishes it immediately (MVP).",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request / Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "requests"
        ],
        "security": [
          {}
        ]
      }
    },
    "/requests/public/{id}": {
      "get": {
        "operationId": "RequestsController_getPublicById",
        "summary": "Get published request by id (public)",
        "description": "Returns a single published request with public client info.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "65f0c1a2b3c4d5e6f7a8b9c1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestPublicDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request / Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "requests"
        ],
        "security": [
          {}
        ]
      }
    },
    "/requests/public": {
      "get": {
        "operationId": "RequestsController_listPublic",
        "summary": "List published requests (for providers)",
        "description": "Returns published requests. Optional filters: lat/lng/radiusKm (nearby), cityId (fallback), categoryKey, subcategoryKey (preferred), serviceKey (deprecated).",
        "parameters": [
          {
            "name": "cityId",
            "required": false,
            "in": "query",
            "schema": {
              "example": "64f0c1a2b3c4d5e6f7a8b9c0",
              "type": "string"
            }
          },
          {
            "name": "lat",
            "required": false,
            "in": "query",
            "description": "Latitude for nearby search (WGS84)",
            "schema": {
              "example": 50.1109,
              "type": "number"
            }
          },
          {
            "name": "lng",
            "required": false,
            "in": "query",
            "description": "Longitude for nearby search (WGS84)",
            "schema": {
              "example": 8.6821,
              "type": "number"
            }
          },
          {
            "name": "radiusKm",
            "required": false,
            "in": "query",
            "description": "Search radius in km. Default 10",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "categoryKey",
            "required": false,
            "in": "query",
            "schema": {
              "example": "cleaning",
              "type": "string"
            }
          },
          {
            "name": "subcategoryKey",
            "required": false,
            "in": "query",
            "schema": {
              "example": "window_cleaning",
              "type": "string"
            }
          },
          {
            "name": "serviceKey",
            "required": false,
            "in": "query",
            "deprecated": true,
            "description": "Use subcategoryKey instead",
            "schema": {
              "example": "home_cleaning",
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "example": "date_desc",
              "enum": [
                "date_desc",
                "date_asc",
                "price_asc",
                "price_desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "priceMin",
            "required": false,
            "in": "query",
            "description": "Filter by price >= priceMin",
            "schema": {
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "priceMax",
            "required": false,
            "in": "query",
            "description": "Filter by price <= priceMax",
            "schema": {
              "example": 200,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (>=1). Default 1",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1..100). Default 20",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Offset for pagination. Default 0",
            "schema": {
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestsPublicResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request / Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "requests"
        ],
        "security": [
          {}
        ]
      }
    },
...
    "/requests/my/photos": {
      "post": {
        "operationId": "RequestsController_uploadMyPhotos",
        "summary": "Client: upload request photos",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Upload up to 8 photos",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "photos": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestPhotosUploadResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request / Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "requests"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/requests/my/{requestId}/publish": {
      "post": {
        "operationId": "RequestsController_publishMy",
        "summary": "Client: publish my draft request",
        "parameters": [
          {
            "name": "requestId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "65f0c1a2b3c4d5e6f7a8b9c1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/RequestResponseDto"
                    },
                    {
                      "example": {
                        "id": "65f0c1a2b3c4d5e6f7a8b9c1",
                        "title": "Zwei IKEA Pax Schränke aufbauen",
                        "serviceKey": "home_cleaning",
                        "cityId": "64f0c1a2b3c4d5e6f7a8b9c0",
                        "cityName": "Frankfurt am Main",
                        "categoryKey": "furniture",
                        "categoryName": "Möbelaufbau",
                        "subcategoryName": "IKEA Aufbau",
                        "propertyType": "apartment",
                        "area": 55,
                        "price": 120,
                        "preferredDate": "2026-02-01T10:00:00.000Z",
                        "isRecurring": false,
                        "comment": "Need eco products, please",
                        "description": "Assemble two wardrobes, tools available",
                        "photos": [
                          "https://cdn.example.com/req/1.jpg"
                        ],
                        "imageUrl": "https://cdn.example.com/req/1.jpg",
                        "tags": [
                          "ikea",
                          "assembly"
                        ],
                        "status": "published",
                        "createdAt": "2026-01-28T10:20:30.123Z"
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request / Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "requests"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Decide API",
    "description": "API documentation",
    "version": "1.0.0",
    "contact": {},
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "https://backend-de-cizhen.onrender.com",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "refreshToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken"
      }
    },
    "schemas": {
...
      "CreateRequestDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Zwei IKEA Pax Schränke aufbauen"
          },
          "serviceKey": {
            "type": "string",
            "example": "home_cleaning"
          },
          "cityId": {
            "type": "string",
            "example": "64f0c1a2b3c4d5e6f7a8b9c0",
            "description": "City id from /catalog/cities"
          },
          "cityName": {
            "type": "string",
            "example": "Frankfurt am Main",
            "description": "City name (used when cityId is not provided)"
          },
          "lat": {
            "type": "number",
            "example": 50.1109,
            "description": "Latitude (WGS84)"
          },
          "lng": {
            "type": "number",
            "example": 8.6821,
            "description": "Longitude (WGS84)"
          },
          "propertyType": {
            "type": "string",
            "example": "apartment",
            "enum": [
              "apartment",
              "house"
            ]
          },
          "area": {
            "type": "number",
            "example": 55,
            "description": "Area in m²"
          },
          "price": {
            "type": "number",
            "example": 120,
            "description": "Optional price in EUR"
          },
          "preferredDate": {
            "type": "string",
            "example": "2026-02-01T10:00:00.000Z"
          },
          "isRecurring": {
            "type": "boolean",
            "example": false
          },
          "comment": {
            "type": "string",
            "example": "Need eco products, please",
            "maxLength": 1000
          },
          "description": {
            "type": "string",
            "example": "Assemble two wardrobes, tools available",
            "maxLength": 2000
          },
          "photos": {
            "example": [
              "https://cdn.example.com/req/1.jpg"
            ],
            "description": "Optional photos for the request card",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tags": {
            "example": [
              "ikea",
              "assembly"
            ],
            "description": "Optional tags for search",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "serviceKey",
          "propertyType",
          "area",
          "preferredDate",
          "isRecurring"
        ]
      },
...
    }
  },
  "security": [
    {
      "access-token": []
    }
  ]
}